Provided riemax.numerical.curves
Providing `riemax.numerical.curves` provides the ability to parameterise
geodesics by cubic splines. These cubic splines themselves are parameterised
by the basis of the null-space of the constraint matrix.

Parameterising geodesics in this manner allows for computing the log map in a
new way, namely: finding parameters of the curve which minimise the energy of
the geodesic. This, along with the shooting method, is the standard approach
for computing the log map in the literature.